requirements.txt
streamlit
pandas
joblib
pdfplumber
langchain
langchain-groq
langchain-huggingface
faiss-cpu
python-dotenv
sentence-transformers
